+++ src/com/xsjiong/vlexer/VNullLexer.java
@@ -2,7 +2,7 @@ package com.xsjiong.vlexer;
 
 public class VNullLexer extends VLexer {
 	@Override
-	protected short getNext() {
+	protected byte getNext() {
 		if (P == L) return VLexer.TYPE_EOF;
 		P = L;
 		return VLexer.UNRESOLVED_TYPE;
